+++ py/picca/utils.py
@@ -81,7 +81,7 @@ def desi_from_truth_to_drq(truth,targets,drq,spectype="QSO"):
     print(" and TRUESPECTYPE=={}  : nb object in cat = {}".format(spectype,w.sum()) )
 
     thid = vac[1]["TARGETID"][:][w]
-    zqso = vac[1]["TRUEZ"][:][w].astype('float64')
+    zqso = vac[1]["TRUEZ"][:][w]
     vac.close()
     ra = sp.zeros(thid.size)
     dec = sp.zeros(thid.size)
@@ -159,7 +159,7 @@ def desi_from_ztarget_to_drq(ztarget,drq,spectype='QSO',downsampling_z_cut=None,
 
     ra = vac[1]['RA'][:][w].astype('float64')
     dec = vac[1]['DEC'][:][w].astype('float64')
-    zqso = vac[1]['Z'][:][w].astype('float64')
+    zqso = vac[1]['Z'][:][w]
     thid = vac[1]['TARGETID'][:][w]
 
     vac.close()
@@ -245,7 +245,7 @@ def desi_convert_transmission_to_delta_files(zcat,indir,outdir,lObs_min=3600.,lO
             continue
         ra = h[1]['RA'][:].astype('float64')*sp.pi/180.
         dec = h[1]['DEC'][:].astype('float64')*sp.pi/180.
-        z = h[1]['Z'][:].astype('float64')
+        z = h[1]['Z'][:]
         ll = sp.log10(h[2].read())
         trans = h[3].read()
         nObj = z.size
